requirements.txt
streamlit
numpy
Pillow
scikit-image
streamlit-image-comparison